:recycle: Auth파트 interface분리
