+++ src/main/java/com/msg/after_school/domain/auth/service/impl/LogoutServiceImpl.java
@@ -1,0 +1,18 @@
+package com.msg.after_school.domain.auth.service.impl;
+
+import com.msg.after_school.domain.auth.service.LogoutService;
+import com.msg.after_school.domain.user.data.entity.User;
+import com.msg.after_school.global.user.dao.impl.UserDaoImpl;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LogoutServiceImpl implements LogoutService {
+    private final UserDaoImpl userDao;
+
+    public void execute() {
+        User user = userDao.getCurrentUser();
+        user.updateRefreshToken(null);
+    }
+}
