+++ src/main/java/com/msg/after_school/global/user/dao/impl/UserDaoImpl.java
@@ -17,7 +17,6 @@ public class UserDaoImpl implements UserDao{
         String email = SecurityContextHolder.getContext().getAuthentication().getName();
         return getUserByEmail(email);
     }
-
     public User getUserByEmail(String email) {
         return userRepository.findUserByEmail(email).orElseThrow(UserNotFoundException::new);
     }
